SYMMETRICDS-562 - Fix DataService bug. Max data id that is currently handled in 2.x is Integer.MAX_VALUE. It should go all the way up to Long.MAX_VALUE

+++ symmetric/symmetric-core/src/main/java/org/jumpmind/symmetric/service/impl/PurgeService.java
@@ -211,8 +211,8 @@ public class PurgeService extends AbstractService implements IPurgeService {
 
         MapSqlParameterSource parameterSource = new MapSqlParameterSource();
         parameterSource.registerSqlType(PARAM_CUTOFF_TIME, Types.TIMESTAMP);
-        parameterSource.registerSqlType(PARAM_MIN, Types.INTEGER);
-        parameterSource.registerSqlType(PARAM_MAX, Types.INTEGER);
+        parameterSource.registerSqlType(PARAM_MIN, Types.NUMERIC);
+        parameterSource.registerSqlType(PARAM_MAX, Types.NUMERIC);
         parameterSource.addValue(PARAM_CUTOFF_TIME, new Timestamp(retentionTime.getTime()));
 
         while (minId <= purgeUpToId) {
